requirements.txt
langchain
chromadb
flask
langchain-community
langchain-chroma
langchain-groq
streamlit_chat
flask_cors